Feature/comprehensive platform update (#37)
* refactor(backend): reorganize DTO structure for better modularity

* feat(backend): implement outbox pattern and SignalR services

* feat(backend): update domain entities, persistence layers and core handlers

* feat(frontend): update api services and state management

* feat(frontend): update UI components and page layouts

* feat(backend): update api controllers and query logic

* chore: general system cleanup and untracked config

* chore(db): add new migrations for posts visibility, comment indexes, media URLs

* refactor(infra): improve Redis caching services and infrastructure resilience

* feat(email): add email service with SMTP, templates, and outbox pattern

* feat(auth): add forgot password, change password, and OTP flow

* feat(api): enhance controllers, add global exception handler and security middleware

* feat(cqrs): add and update command/query handlers across all domains

* feat(frontend): add forgot/reset password UI, enhance questions, groups, settings pages

* test: add integration tests, k6 load tests, playwright e2e tests, and production deployment configs

* chore: add appsettings.json to gitignore to prevent secrets exposure

* feat(api): track appsettings.json with non-sensitive config

Move sensitive values to user-secrets for security.

+++ backend/src/SocialTechsy.SocialNetwork.Api/Hubs/ChatHub.cs
@@ -254,7 +254,7 @@ public class ChatHub : Hub
             .SendAsync("RemoveReaction", new { messageId, userId });
     }
 
-    public async Task EditMessage(int conversationId, int messageId, string newContent)
+    public async Task EditMessage(int conversationId, long messageId, string newContent)
     {
         var userId = GetCurrentUserId();
         if (userId == 0) return;
@@ -278,7 +278,7 @@ public class ChatHub : Hub
             .SendAsync("MessageEdited", new { messageId, newContent, editedDate = DateTime.UtcNow });
     }
 
-    public async Task DeleteMessage(int conversationId, int messageId)
+    public async Task DeleteMessage(int conversationId, long messageId)
     {
         var userId = GetCurrentUserId();
         if (userId == 0) return;
